chore(core): keep resolveVersionRelease internal to hooks package
