+++ packages/sanity/src/core/hooks/index.ts
@@ -30,5 +30,5 @@ export * from './useTools'
 export * from './useUnitFormatter'
 export * from './useUserListWithPermissions'
 export * from './useValidationStatus'
-export * from './useVersionRelease'
+export {useVersionRelease} from './useVersionRelease'
 export * from './useWorkspaceSchemaId'
